add user restrict and fix order page bugs

+++ routes/web.php
@@ -25,19 +25,21 @@ use App\Http\Controllers\TransactionController;
 
 Auth::routes();
 
+// dd(User::where('id', ));
+// $dashboard = ( Auth::user() ->status == 1 ) ? 'adminHome' : 'home';
+
 //Base Pages
 Route::get('/', [HomeController::class, 'index'])->name('home');
 Route::get('/home', [HomeController::class, 'index'])->name('home');
-Route::get('admin/home', [HomeController::class, 'adminHome'])->name('admin.home')->middleware('is_admin');
+Route::get('admin/home', [HomeController::class, 'index'])->name('admin.home') ->middleware('is_admin');
 
 //Comic Pages
 Route::get('/store', [BooksController::class, 'index'])->name('comic');
-Route::post('/comic/save', [BooksController::class, 'store'])->name('comic');
-Route::get('/comic/detail/{id}', [BooksController::class, 'show'])->name('comic');
-Route::get('/comic/edit/{id}', [BooksController::class, 'edit'])->name('comic');
-Route::delete('/comic/delete/{id}', [BooksController::class, 'destroy'])->name('comic');
-Route::post('/comic/update/{id}', [BooksController::class, 'update'])->name('comic');
-
+Route::post('/comic/save', [BooksController::class, 'store'])->name('comic')  ->middleware('is_admin');
+Route::get('/comic/detail/{id}', [BooksController::class, 'show'])->name('comic')  ->middleware('is_admin');
+Route::get('/comic/edit/{id}', [BooksController::class, 'edit'])->name('comic')  ->middleware('is_admin');
+Route::delete('/comic/delete/{id}', [BooksController::class, 'destroy'])->name('comic')  ->middleware('is_admin');
+Route::post('/comic/update/{id}', [BooksController::class, 'update'])->name('comic')  ->middleware('is_admin');
 
 //Order Pages
 Route::get('/order/{id}', [OrdersController::class, 'index'])-> where('id', '[0-9]+');
@@ -47,16 +49,16 @@ Route::delete('/checkout/remove/{id}', [OrdersController::class, 'remove']) ->na
 Route::get('/checkout/confirm', [OrdersController::class, 'checkout_confirm']) ->name('checkout_confirm');
 
 //Transaction pages
-Route::get('/transactions/list', [TransactionController::class, 'index']) ->name('transactions');
-Route::get('/transactions/details/{id}', [TransactionController::class, 'show']) ->name('transactions');
-Route::get('/transactions/edit/{id}', [TransactionController::class, 'edit']) ->name('transactions');
-Route::post('/transactions/update/{id}', [TransactionController::class, 'update']) ->name('transactions');
+Route::get('/transactions/list', [TransactionController::class, 'index']) ->name('transactions')  ->middleware('is_admin');
+Route::get('/transactions/details/{id}', [TransactionController::class, 'show']) ->name('transactions')  ->middleware('is_admin');
+Route::get('/transactions/edit/{id}', [TransactionController::class, 'edit']) ->name('transactions')  ->middleware('is_admin');
+Route::post('/transactions/update/{id}', [TransactionController::class, 'update']) ->name('transactions')  ->middleware('is_admin');
 
 //User Pages
-Route::get('/users/list', [UsersController::class, 'index'])->name('users');
-Route::get('/user/history/{id}', [UsersController::class, 'history'])->name('users');
-Route::get('/user/details/{id}', [UsersController::class, 'show'])->name('users');
-Route::delete('/user/delete/{id}', [UsersController::class, 'destroy'])->name('users');
+Route::get('/users/list', [UsersController::class, 'index'])->name('users') ->middleware('is_admin');
+Route::get('/user/history/{id}', [UsersController::class, 'history'])->name('users') ->middleware('is_admin');
+Route::get('/user/details/{id}', [UsersController::class, 'show'])->name('users') ->middleware('is_admin');
+Route::delete('/user/delete/{id}', [UsersController::class, 'destroy'])->name('users') ->middleware('is_admin');
 
 
 //User Guest Pages
@@ -69,7 +71,7 @@ Route::get('/history', [HistoryController::class, 'index'])->name('history');
 Route::get('/history/{id}', [HistoryController::class, 'details'])->name('history');
 
 //Admin Page
-Route::get('/comics/list', [BooksController::class, 'bookList'])->name('comic');
+Route::get('/comics/list', [BooksController::class, 'bookList'])->name('comic')  ->middleware('is_admin');;
 
 
 Route::group(['middleware' => 'auth'], function () {
